// dotnetydd-vue-template/dotnetydd-vadmin/src/utils/http/types.d.ts
import type {
    AxiosError,
    AxiosRequestConfig,
    AxiosResponse,
    Method
} from "axios";

export type RequestMethods = Extract<
    Method,
    "get" | "post" | "put" | "delete" | "patch" | "option" | "head"
>;


export interface PureHttpRequestConfig extends AxiosRequestConfig {
    beforeRequestCallback?: (request: PureHttpRequestConfig) => void;
    beforeResponseCallback?: (response: PureHttpResponse) => void;
}

export interface PureHttpError extends AxiosError {
    isCancelRequest?: boolean;
}

export interface PureHttpResponse extends AxiosResponse {
    config: PureHttpRequestConfig;
}


export default class PureHttp {
    request<T>(
        method: RequestMethods,
        url: string,
        param?: AxiosRequestConfig,
        axiosConfig?: PureHttpRequestConfig
    ): Promise<T>;
    post<T, P>(
        url: string,
        params?: P,
        config?: PureHttpRequestConfig
    ): Promise<T>;
    get<T, P>(
        url: string,
        params?: P,
        config?: PureHttpRequestConfig
    ): Promise<T>;
}